RungeKuttaToolKit.jl

RungeKuttaToolKit.RKOCEvaluator — Type

RKOCEvaluator{T}(
    trees::AbstractVector{LevelSequence},
    num_stages::Int;
    optimize::Bool=true,
    sort_by_depth::Bool=true,
) -> RKOCEvaluator{T}

Construct an RKOCEvaluator that encodes a given sequence of rooted trees.

Arguments

trees: input vector of rooted trees in LevelSequence representation.

num_stages: number of stages (i.e., size of the Butcher tableau). Must be known at construction time to allocate internal workspace arrays.

optimize: if true, perform common subtree elimination. This may improve performance in cases where trees is not the complete set of all rooted trees up to a certain order.

sort_by_depth: if true, permute internal workspace arrays so that intermediate results are calculated in an order that enables parallel execution. This has no effect on single-threaded execution and is provided for forward compatibility with future parallel implementations.

Note that different permutations of trees, in addition to different values of optimize and sort_by_depth, may yield slightly different results due to the non-associative nature of floating-point arithmetic.

RKOCEvaluator{T}(order::Int, num_stages::Int) -> RKOCEvaluator{T}

Construct an RKOCEvaluator that encodes all rooted trees having at most order vertices.

By default, rooted trees are generated in graded reverse lexicographic order of their level sequence representation. This specific ordering maximizes the efficiency of generating all rooted trees.

(ev::RKOCEvaluator{T})(
    residuals::AbstractVector{T},
    A::AbstractMatrix{T},
    b::AbstractVector{T}
) -> AbstractVector{T}

Compute residuals of the Runge–Kutta order conditions $\{ \mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t) : t \in T \}$ for a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ encoded by an RKOCEvaluator.

Arguments

ev: RKOCEvaluator object encoding a set of rooted trees.

residuals: length $|T|$ output vector. Each residual $\mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t)$ is written to residuals[i] in the order specified when constructing ev.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

(ev::RKOCEvaluator{T})(
    A::AbstractMatrix{T},
    b::AbstractVector{T}
) -> T

Compute the sum of squared residuals of the Runge–Kutta order conditions $\sum_{t \in T} (\mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t))^2$ for a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ encoded by an RKOCEvaluator.

Arguments

ev: RKOCEvaluator object encoding a set of rooted trees.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $s$ denotes the number of stages specified when constructing ev.

RungeKuttaToolKit.WeightedRKOCEvaluator — Type

WeightedRKOCEvaluator{T}(
    trees::AbstractVector{LevelSequence},
    weights::AbstractVector{T},
    num_stages::Int;
    optimize::Bool=true,
    sort_by_depth::Bool=true,
) -> WeightedRKOCEvaluator{T}

Construct a WeightedRKOCEvaluator that encodes a given sequence $T$ of rooted trees $t \in T$ with associated weights $w_t$.

Arguments

trees: input vector of rooted trees in LevelSequence representation.

weights: input vector of weights associated with each rooted tree. Must have the same shape as trees.

num_stages: number of stages (i.e., size of the Butcher tableau). Must be known at construction time to allocate internal workspace arrays.

optimize: if true, perform common subtree elimination. This may improve performance in cases where trees is not the complete set of all rooted trees up to a certain order.

sort_by_depth: if true, permute internal workspace arrays so that intermediate results are calculated in an order that enables parallel execution. This has no effect on single-threaded execution and is provided for forward compatibility with future parallel implementations.

Note that different permutations of trees, in addition to different values of optimize and sort_by_depth, may yield slightly different results due to the non-associative nature of floating-point arithmetic.

WeightedRKOCEvaluator{T}(
    order::Int,
    num_stages::Int,
) -> WeightedRKOCEvaluator{T}

Construct a WeightedRKOCEvaluator that encodes all rooted trees $t$ having at most order vertices with associated weights $w_t = 1/\sigma(t)$. Under this choice of weight, the residuals the become principal error coefficients of a Runge–Kutta method.

By default, rooted trees are generated in graded reverse lexicographic order of their level sequence representation. This specific ordering maximizes the efficiency of generating all rooted trees.

(ev::WeightedRKOCEvaluator{T})(
    residuals::AbstractVector{T},
    A::AbstractMatrix{T},
    b::AbstractVector{T}
) -> AbstractVector{T}

Compute weighted residuals of the Runge–Kutta order conditions $\{ w_t (\mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t)) : t \in T \}$ for a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ with associated weights $w_t$.

(ev::WeightedRKOCEvaluator{T})(
    A::AbstractMatrix{T},
    b::AbstractVector{T}
) -> T

Compute the weighted sum of squared residuals of the Runge–Kutta order conditions $\sum_{t \in T} w_t^2 (\mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t))^2$ for a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ with associated weights $w_t$.

RungeKuttaToolKit.RKOCAdjoint — Type

(adj::RKOCAdjoint{T})(
    dresiduals::AbstractVector{T},
    A::AbstractMatrix{T},
    dA::AbstractMatrix{T},
    b::AbstractVector{T},
    db::AbstractVector{T},
) -> AbstractVector{T}

Compute directional derivatives of the Runge–Kutta order conditions $\{ \nabla_{\mathrm{d}A, \mathrm{d}\mathbf{b}} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ at a given Butcher tableau $(A, \mathbf{b})$ in direction $(\mathrm{d}A, \mathrm{d}\mathbf{b})$ over a set of rooted trees $T$ encoded by an RKOCEvaluator.

Arguments

adj: RKOCAdjoint object obtained by applying the adjoint operator ' to an RKOCEvaluator. In other words, this function should be called as ev'(dresiduals, A, dA, b, db) where ev is an RKOCEvaluator.

dresiduals: length $|T|$ output vector. Each directional derivative $\nabla_{\mathrm{d}A, \mathrm{d}\mathbf{b}} [ \mathbf{b} \cdot \Phi_t(A) ]$ is written to dresiduals[i] in the order specified when constructing ev.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

dA: $s \times s$ input matrix containing the direction in which to differentiate $A$.

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

db: length $s$ input vector containing the direction in which to differentiate $\mathbf{b}$.

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

(adj::RKOCAdjoint{T})(
    dresiduals::AbstractVector{T},
    A::AbstractMatrix{T},
    i::Int,
    j::Int,
    b::AbstractVector{T},
) -> AbstractVector{T}

Compute partial derivatives of the Runge–Kutta order conditions $\{ \partial_{A_{i,j}} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ with respect to $A_{i,j}$ at a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ encoded by an RKOCEvaluator.

Arguments

adj: RKOCAdjoint object obtained by applying the adjoint operator ' to an RKOCEvaluator. In other words, this function should be called as ev'(dresiduals, A, i, j, b) where ev is an RKOCEvaluator.

dresiduals: length $|T|$ output vector. Each partial derivative $\partial_{A_{i,j}} [ \mathbf{b} \cdot \Phi_t(A) ]$ is written to dresiduals[i] in the order specified when constructing ev.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

i: row index of the entry of $A$ to differentiate with respect to.

j: column index of the entry of $A$ to differentiate with respect to.

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

This method uses an optimized algorithm that should be faster than ev'(dresiduals, A, dA, b, db) when differentiating with respect to a single entry of $A$. It may produce slightly different results due to the non-associative nature of floating-point arithmetic.

(adj::RKOCAdjoint{T})(
    dresiduals::AbstractVector{T},
    A::AbstractMatrix{T},
    i::Int,
) -> AbstractVector{T}

Compute partial derivatives of the Runge–Kutta order conditions $\{ \partial_{b_i} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ with respect to $b_i$ at a given Butcher tableau $A$ over a set of rooted trees $T$ encoded by an RKOCEvaluator. (The result is independent of the value of $\mathbf{b}$.)

Arguments

adj: RKOCAdjoint object obtained by applying the adjoint operator ' to an RKOCEvaluator. In other words, this function should be called as ev'(dresiduals, A, i) where ev is an RKOCEvaluator.

dresiduals: length $|T|$ output vector. Each partial derivative $\partial_{b_i} [ \mathbf{b} \cdot \Phi_t(A) ]$ is written to dresiduals[i] in the order specified when constructing ev.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

i: index of the entry of $\mathbf{b}$ to differentiate with respect to.

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

This method uses an optimized algorithm that should be faster than ev'(dresiduals, A, dA, b, db) when differentiating with respect to a single entry of $\mathbf{b}$. It may produce slightly different results due to the non-associative nature of floating-point arithmetic.

(adj::RKOCAdjoint{T})(
    dA::AbstractMatrix{T},
    db::AbstractVector{T},
    A::AbstractMatrix{T},
    b::AbstractVector{T},
) -> Tuple{AbstractMatrix{T}, AbstractVector{T}}

Compute the gradient of the sum of squared residuals of the Runge–Kutta order conditions $\nabla_{A, \mathbf{b}} \sum_{t \in T} (\mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t))^2$ at a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ encoded by an RKOCEvaluator.

Arguments

adj: RKOCAdjoint object obtained by applying the adjoint operator ' to an RKOCEvaluator. In other words, this function should be called as ev'(dA, db, A, b) where ev is an RKOCEvaluator.

dA: $s \times s$ output matrix containing the gradient of the sum of squared residuals with respect to $A$.

db: length $s$ output vector containing the gradient of the sum of squared residuals with respect to $\mathbf{b}$.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $s$ denotes the number of stages specified when constructing ev.

RungeKuttaToolKit.WeightedRKOCAdjoint — Type

(adj::WeightedRKOCAdjoint{T})(
    dresiduals::AbstractVector{T},
    A::AbstractMatrix{T},
    dA::AbstractMatrix{T},
    b::AbstractVector{T},
    db::AbstractVector{T},
) -> AbstractVector{T}

Compute weighted directional derivatives of the Runge–Kutta order conditions $\{ w_t \nabla_{\mathrm{d}A, \mathrm{d}\mathbf{b}} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ at a given Butcher tableau $(A, \mathbf{b})$ in direction $(\mathrm{d}A, \mathrm{d}\mathbf{b})$ over a set of rooted trees $T$ with associated weights $w_t$.

(adj::WeightedRKOCAdjoint{T})(
    dresiduals::AbstractVector{T},
    A::AbstractMatrix{T},
    i::Int,
    j::Int,
    b::AbstractVector{T},
) -> AbstractVector{T}

Compute weighted partial derivatives of the Runge–Kutta order conditions $\{ w_t \partial_{A_{i,j}} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ with respect to $A_{i,j}$ at a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ with associated weights $w_t$.

(adj::WeightedRKOCAdjoint{T})(
    dresiduals::AbstractVector{T},
    A::AbstractMatrix{T},
    i::Int,
) -> AbstractVector{T}

Compute weighted partial derivatives of the Runge–Kutta order conditions $\{ w_t \partial_{b_i} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ with respect to $b_i$ at a given Butcher tableau $A$ over a set of rooted trees $T$ with associated weights $w_t$. (The result is independent of the value of $\mathbf{b}$.)

(adj::WeightedRKOCAdjoint{T})(
    dA::AbstractMatrix{T},
    db::AbstractVector{T},
    A::AbstractMatrix{T},
    b::AbstractVector{T},
) -> Tuple{AbstractMatrix{T}, AbstractVector{T}}

Compute the gradient of the weighted sum of squared residuals of the Runge–Kutta order conditions $\nabla_{A, \mathbf{b}} \sum_{t \in T} w_t^2 (\mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t))^2$ at a given Butcher tableau $(A, \mathbf{b})$ over a set of rooted trees $T$ with associated weights $w_t$.