RungeKuttaToolKit.jl

RungeKuttaToolKit.jl is a Julia package for designing and analyzing Runge–Kutta methods for solving ordinary differential equations. The main tool it provides is the RKOCEvaluator type, which implements fast algorithms for evaluating the residuals and gradients of the Runge–Kutta order conditions

\[\mathbf{b} \cdot \Phi_t(A) = \frac{1}{\gamma(t)}\]

over a given set of rooted trees $t \in T$.

Features

RungeKuttaToolKit.jl is designed for extremely high performance and stable memory usage. It evaluates Runge–Kutta order conditions and their derivatives millions of times per second on a single CPU core, using fast in-place algorithms that allocate zero memory outside of RKOCEvaluator construction.

RungeKuttaToolKit.jl works with all numeric types that support arithmetic (+ - * /), including low-precision and high-precision floating-point types, BigFloat, Rational, and even symbolic types.

RungeKuttaToolKit.jl is lightweight and has zero dependencies outside of Julia Base.

RungeKuttaToolKit.jl is regularly and extensively tested for correctness and performance.

Quick Start

To install RungeKuttaToolKit.jl, run ]add RungeKuttaToolKit in the Julia REPL.

using RungeKuttaToolKit

order = 4
num_stages = 4

# RK4 Butcher tableau
A = [0.0 0.0 0.0 0.0;
     0.5 0.0 0.0 0.0;
     0.0 0.5 0.0 0.0;
     0.0 0.0 1.0 0.0]
b = [1/6, 1/3, 1/3, 1/6]

ev = RKOCEvaluator(order, num_stages)
residuals = ev(A, b)
println(residuals) # This should be close to zero,
# which confirms that RK4 is a fourth-order method.

# An RKOCEvaluator can also be constructed from a set of
# rooted trees. Here, we select only the trees of order 5.
trees = rooted_trees(order + 1)
sigma = inv.(Float64.(butcher_symmetry.(trees)))
err_ev = RKOCEvaluator(trees, num_stages)
error_coeffs = sigma .* err_ev(A, b)
println(error_coeffs) # This computes the principal error
# coefficients of RK4, which should all be nonzero.

Generating Rooted Trees

The first step in using RungeKuttaToolKit.jl is generating a list of rooted trees. Each Runge–Kutta order condition $\mathbf{b} \cdot \Phi_t(A) = 1/\gamma(t)$ corresponds to a rooted tree $t$. For a Runge–Kutta method $(A, \mathbf{b})$ to have order $p$, it needs to satisfy this equation for all rooted trees $t$ having up to $p$ vertices.

RungeKuttaToolKit.jl provides two functions for generating rooted trees: rooted_trees(n), which generates all rooted trees with exactly $n$ vertices, and all_rooted_trees(n), which generates all rooted trees with at most $n$ vertices.

RungeKuttaToolKit.ButcherInstructions.rooted_trees — Function

rooted_trees(
    n::Integer;
    tree_ordering::Symbol=:reverse_lexicographic,
) -> Vector{LevelSequence}

Compute and return a list of all rooted trees with exactly $n$ vertices. Several algorithms that generate rooted trees in different orderings are available:

:reverse_lexicographic (default) is the asymptotically optimal algorithm devised by Terry Beyer and Sandra Mitchell Hedetniemi that generates rooted trees in constant time per tree. It generates rooted trees in reverse lexicographic order of their canonical level sequence representation. This is the fastest known algorithm, but unlike the others, it produces the tall tree first and the bushy tree last.

:lexicographic generates rooted trees in lexicographic order of their canonical level sequence representation. It simply runs the :reverse_lexicographic algorithm and reverses its output. This algorithm produces the bushy tree first and the tall tree last.

:attach generates rooted trees of order $n$ by attaching a leaf to each vertex of each rooted tree of order $n-1$, starting from the root and proceeding depth-first. This is the ordering produced by paper-and-pencil application of the product rule to the set of elementary differentials. This is the slowest algorithm, but it is arguably the most intuitive. It produces the bushy tree first and the tall tree last.

:butcher is Algorithm 3 from John Butcher's monograph on B-Series, which generates rooted trees in a canonical ordering often used in the literature on Runge–Kutta methods. It produces the bushy tree first and the tall tree last.

RungeKuttaToolKit.ButcherInstructions.all_rooted_trees — Function

all_rooted_trees(
    n::Integer;
    tree_ordering::Symbol=:reverse_lexicographic,
) -> Vector{LevelSequence}

Compute and return a list of all rooted trees with up to $n$ vertices. Several algorithms that generate rooted trees in different orderings are available:

:reverse_lexicographic (default) is the asymptotically optimal algorithm devised by Terry Beyer and Sandra Mitchell Hedetniemi that generates rooted trees in constant time per tree. It generates rooted trees in reverse lexicographic order of their canonical level sequence representation. This is the fastest known algorithm, but unlike the others, it produces the tall tree first and the bushy tree last.

:lexicographic generates rooted trees in lexicographic order of their canonical level sequence representation. It simply runs the :reverse_lexicographic algorithm and reverses its output. This algorithm produces the bushy tree first and the tall tree last.

:attach generates rooted trees of order $n$ by attaching a leaf to each vertex of each rooted tree of order $n-1$, starting from the root and proceeding depth-first. This is the ordering produced by paper-and-pencil application of the product rule to the set of elementary differentials. This is the slowest algorithm, but it is arguably the most intuitive. It produces the bushy tree first and the tall tree last.

:butcher is Algorithm 3 from John Butcher's monograph on B-Series, which generates rooted trees in a canonical ordering often used in the literature on Runge–Kutta methods. It produces the bushy tree first and the tall tree last.

RKOCEvaluator Construction

With a list of rooted trees in hand, we can construct an RKOCEvaluator object, which is used to evaluate the residuals and gradients of the Runge–Kutta order conditions with respect to the entries of $A$ and $\mathbf{b}$. The RKOCEvaluator constructor analyzes the given list of rooted trees to identify common subtrees whose Butcher weight vectors $\Phi_t(A)$ can be reused to eliminate redundant computation.

RungeKuttaToolKit.RKOCEvaluator — Type

RKOCEvaluator{T}(
    trees::AbstractVector{LevelSequence},
    num_stages::Integer,
) -> RKOCEvaluator{T}

Construct an RKOCEvaluator that encodes a given list of rooted trees.

Arguments

trees: input list of rooted trees in LevelSequence representation.

num_stages: number of stages (i.e., size of the Butcher tableau). Must be specified at construction time to allocate internal workspace arrays.

If the type T is not specified, it defaults to Float64.

RKOCEvaluator{T}(order::Integer, num_stages::Integer) -> RKOCEvaluator{T}

Construct an RKOCEvaluator that encodes all rooted trees having at most order vertices. This is equivalent to RKOCEvaluator{T}(all_rooted_trees(order), num_stages).

If the type T is not specified, it defaults to Float64.

RKOCEvaluator Usage

An RKOCEvaluator object can be used to perform seven operations which are summarized below. A full description of each operation follows the summary.

ev([residuals,] A, b): Compute residuals.

ev(cost, A, b): Compute cost function.

ev'([dresiduals,] A, dA, b, db): Compute directional derivatives. of residuals

ev'([dresiduals,] A, i, j, b): Compute partial derivatives of residuals. with respect to $A_{i,j}$

ev'([dresiduals,] A, i): Compute partial derivatives of residuals. with respect to $b_i$

ev'([dA, db,] cost, A, b): Compute gradient of cost function.

ev'([jacobian, A, b,] param, x): Compute Jacobian of residuals with respect to parameterization.

Every operation that returns a matrix or vector has an allocating version that returns its result in a new array and an in-place version that writes its result to an existing array. We use square brackets to denote optional output arguments. For example, the operation ev([residuals,] A, b) can be called as ev(A, b) (allocating) or ev(residuals, A, b) (in-place).

As a mnemonic device, the operations that compute derivatives are called with a single quote ' following the RKOCEvaluator object, emulating the prime symbol $f'(x)$ used to denote derivatives in mathematical notation. This distinguishes operations like ev(cost, A, b) and ev'(cost, A, b) that would otherwise have identical function signatures.

RungeKuttaToolKit.AbstractRKOCEvaluator — Type

(ev::AbstractRKOCEvaluator{T})(
    [residuals::AbstractVector{T},]
    A::AbstractMatrix{T},
    b::AbstractVector{T},
) -> AbstractVector{T}

Compute residuals of the Runge–Kutta order conditions $\{ \mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t) : t \in T \}$ for a given Butcher tableau $(A, \mathbf{b})$ over a list of rooted trees $T$ encoded by an AbstractRKOCEvaluator.

Arguments

ev: AbstractRKOCEvaluator object encoding a list of rooted trees.

residuals: length $|T|$ output vector. Each residual $\mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t)$ is written to residuals[i] in the order specified when constructing ev. If not provided, a new vector is allocated and returned.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

(ev::AbstractRKOCEvaluator{T})(
    cost::AbstractRKCost{T},
    A::AbstractMatrix{T},
    b::AbstractVector{T},
) -> T

Compute a specified cost function $f$ of the residuals of the Runge–Kutta order conditions $f(\{ \mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t) : t \in T \})$ for a given Butcher tableau $(A, \mathbf{b})$ over a list of rooted trees $T$ encoded by an AbstractRKOCEvaluator.

Arguments

ev: AbstractRKOCEvaluator object encoding a list of rooted trees.

cost: AbstractRKCost object specifying the cost function $f$.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $s$ denotes the number of stages specified when constructing ev.

RungeKuttaToolKit.AbstractRKOCAdjoint — Type

(adj::AbstractRKOCAdjoint{T})(
    [dresiduals::AbstractVector{T},]
    A::AbstractMatrix{T},
    dA::AbstractMatrix{T},
    b::AbstractVector{T},
    db::AbstractVector{T},
) -> AbstractVector{T}

Compute directional derivatives of the Runge–Kutta order conditions $\{ \nabla_{\mathrm{d}A, \mathrm{d}\mathbf{b}} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ at a given Butcher tableau $(A, \mathbf{b})$ in direction $(\mathrm{d}A, \mathrm{d}\mathbf{b})$ over a list of rooted trees $T$ encoded by an AbstractRKOCEvaluator.

Arguments

adj: AbstractRKOCAdjoint object obtained by applying the adjoint operator ' to an AbstractRKOCEvaluator. In other words, this function should be called as ev'([dresiduals,] A, dA, b, db) where ev is an AbstractRKOCEvaluator.

dresiduals: length $|T|$ output vector. Each directional derivative $\nabla_{\mathrm{d}A, \mathrm{d}\mathbf{b}} [ \mathbf{b} \cdot \Phi_t(A) ]$ is written to dresiduals[i] in the order specified when constructing ev. If not provided, a new vector is allocated and returned.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

dA: $s \times s$ input matrix containing the direction in which to differentiate $A$.

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

db: length $s$ input vector containing the direction in which to differentiate $\mathbf{b}$.

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

(adj::AbstractRKOCAdjoint{T})(
    [dresiduals::AbstractVector{T},]
    A::AbstractMatrix{T},
    i::Integer,
    j::Integer,
    b::AbstractVector{T},
) -> AbstractVector{T}

Compute partial derivatives of the Runge–Kutta order conditions $\{ \partial_{A_{i,j}} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ with respect to $A_{i,j}$ at a given Butcher tableau $(A, \mathbf{b})$ over a list of rooted trees $T$ encoded by an AbstractRKOCEvaluator.

Arguments

adj: AbstractRKOCAdjoint object obtained by applying the adjoint operator ' to an AbstractRKOCEvaluator. In other words, this function should be called as ev'([dresiduals,] A, dA, b, db) where ev is an AbstractRKOCEvaluator.

dresiduals: length $|T|$ output vector. Each partial derivative $\partial_{A_{i,j}} [ \mathbf{b} \cdot \Phi_t(A) ]$ is written to dresiduals[i] in the order specified when constructing ev. If not provided, a new vector is allocated and returned.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

i: row index of the entry of $A$ to differentiate with respect to.

j: column index of the entry of $A$ to differentiate with respect to.

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

This method uses an optimized algorithm that should be faster than ev'(dresiduals, A, dA, b, db) when differentiating with respect to a single entry of $A$. It may produce slightly different results due to the non-associative nature of floating-point arithmetic.

(adj::AbstractRKOCAdjoint{T})(
    [dresiduals::AbstractVector{T},]
    A::AbstractMatrix{T},
    i::Integer,
) -> AbstractVector{T}

Compute partial derivatives of the Runge–Kutta order conditions $\{ \partial_{b_i} [ \mathbf{b} \cdot \Phi_t(A) ] : t \in T \}$ with respect to $b_i$ at a given Butcher tableau $A$ over a list of rooted trees $T$ encoded by an AbstractRKOCEvaluator. (The result is independent of the value of $\mathbf{b}$.)

Arguments

adj: AbstractRKOCAdjoint object obtained by applying the adjoint operator ' to an AbstractRKOCEvaluator. In other words, this function should be called as ev'([dresiduals,] A, dA, b, db) where ev is an AbstractRKOCEvaluator.

dresiduals: length $|T|$ output vector. Each partial derivative $\partial_{b_i} [ \mathbf{b} \cdot \Phi_t(A) ]$ is written to dresiduals[i] in the order specified when constructing ev. If not provided, a new vector is allocated and returned.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

i: index of the entry of $\mathbf{b}$ to differentiate with respect to.

Here, $|T|$ denotes the number of rooted trees encoded by ev, and $s$ denotes the number of stages specified when constructing ev.

This method uses an optimized algorithm that should be faster than ev'(dresiduals, A, dA, b, db) when differentiating with respect to a single entry of $\mathbf{b}$. It may produce slightly different results due to the non-associative nature of floating-point arithmetic.

(adj::AbstractRKOCAdjoint{T})(
    [dA::AbstractMatrix{T},
     db::AbstractVector{T},]
    cost::AbstractRKCost{T},
    A::AbstractMatrix{T},
    b::AbstractVector{T},
) -> Tuple{AbstractMatrix{T}, AbstractVector{T}}

Compute the gradient of a specified cost function $f$ of the residuals of the Runge–Kutta order conditions $\nabla_{A, \mathbf{b}} f(\{ \mathbf{b} \cdot \Phi_t(A) - 1/\gamma(t) : t \in T \})$ at a given Butcher tableau $(A, \mathbf{b})$ over a list of rooted trees $T$ encoded by an AbstractRKOCEvaluator.

Arguments

adj: AbstractRKOCAdjoint object obtained by applying the adjoint operator ' to an AbstractRKOCEvaluator. In other words, this function should be called as ev'([dA, db,] cost, A, b) where ev is an AbstractRKOCEvaluator.

dA: $s \times s$ output matrix containing the gradient of $f$ with respect to $A$. If not provided, a new matrix is allocated and returned.

db: length $s$ output vector containing the gradient of $f$ with respect to $\mathbf{b}$. If not provided, a new vector is allocated and returned.

cost: AbstractRKCost object specifying the cost function $f$.

A: $s \times s$ input matrix containing the coefficients of a Runge–Kutta method (i.e., the upper-right block of a Butcher tableau).

b: length $s$ input vector containing the weights of a Runge–Kutta method (i.e., the lower-right row of a Butcher tableau).

Here, $s$ denotes the number of stages specified when constructing ev.

(adj::AbstractRKOCAdjoint{T})(
    [jacobian::AbstractMatrix{T},
     A::AbstractMatrix{T},
     b::AbstractVector{T},]
    param::AbstractRKParameterization{T},
    x::AbstractVector{T},
) -> AbstractMatrix{T}

Compute the Jacobian of the residuals of the Runge–Kutta order conditions $\nabla_{\mathbf{x}} \{ \mathbf{b}(\mathbf{x}) \cdot \Phi_t(A(\mathbf{x})) - 1/\gamma(t) : t \in T \}$ with respect to a specified Butcher tableau parameterization $\mathbf{x} \in \mathbb{R}^n \mapsto (A(\mathbf{x}), \mathbf{b}(\mathbf{x}))$ over a list of rooted trees $T$ encoded by an AbstractRKOCEvaluator.

Arguments

adj: AbstractRKOCAdjoint object obtained by applying the adjoint operator ' to an AbstractRKOCEvaluator. In other words, this function should be called as ev'([jacobian, A, b,] param, x) where ev is an AbstractRKOCEvaluator.

jacobian: $|T| \times n$ output matrix. The partial derivative of the $i$th residual with respect to the $j$th coordinate of $\mathbf{x}$ is written to jacobian[i, j]. If not provided, a new matrix is allocated and returned.

A: $s \times s$ output matrix containing the output $A(\mathbf{x})$ of the Butcher tableau parameterization. If not provided, a new temporary matrix is allocated.

b: length $s$ output vector containing the output $\mathbf{b}(\mathbf{x})$ of the Butcher tableau parameterization. If not provided, a new temporary vector is allocated.

param: AbstractRKParameterization object specifying a Butcher tableau parameterization, i.e., a mapping from a vector of parameters $\mathbf{x} \in \mathbb{R}^n$ to a Butcher tableau $(A(\mathbf{x}), \mathbf{b}(\mathbf{x}))$.

x: length $n$ input vector containing the parameters $\mathbf{x}$.

Here, $|T|$ denotes the number of rooted trees encoded by ev, $s$ denotes the number of stages specified when constructing ev, and $n$ denotes the number of free parameters in param.

Cost Functions

RungeKuttaToolKit.RKCost.RKCostL1 — Type

\[\sum_{t \in T} \left\lvert \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right\rvert\]

RungeKuttaToolKit.RKCost.RKCostWeightedL1 — Type

\[\sum_{t \in T} w_t \left\lvert \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right\rvert\]

RungeKuttaToolKit.RKCost.RKCostL2 — Type

\[\sum_{t \in T} \left( \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right)^2\]

RungeKuttaToolKit.RKCost.RKCostWeightedL2 — Type

\[\sum_{t \in T} w_t \left( \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right)^2\]

RungeKuttaToolKit.RKCost.RKCostLInfinity — Type

\[\max_{t \in T} \left\lvert \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right\rvert\]

Derivatives are not yet implemented for RKCostLInfinity.

RungeKuttaToolKit.RKCost.RKCostWeightedLInfinity — Type

\[\max_{t \in T} w_t \left\lvert \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right\rvert\]

Derivatives are not yet implemented for RKCostWeightedLInfinity.

RungeKuttaToolKit.RKCost.RKCostHuber — Type

\[\sum_{t \in T} H_\delta\!\left( \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right)\]

\[H_\delta(x) = \begin{cases} x^2 & \text{if } \lvert x \rvert \leq \delta \\ 2 \delta \lvert x \rvert - \delta^2 & \text{otherwise} \end{cases}\]

RungeKuttaToolKit.RKCost.RKCostWeightedHuber — Type

\[\sum_{t \in T} w_t H_\delta\!\left( \mathbf{b} \cdot \Phi_t(A) - \frac{1}{\gamma(t)} \right)\]

\[H_\delta(x) = \begin{cases} x^2 & \text{if } \lvert x \rvert \leq \delta \\ 2 \delta \lvert x \rvert - \delta^2 & \text{otherwise} \end{cases}\]

Parameterizations

RungeKuttaToolKit.RKParameterization.RKParameterizationExplicit — Type

RungeKuttaToolKit.RKParameterization.RKParameterizationDiagonallyImplicit — Type

RungeKuttaToolKit.RKParameterization.RKParameterizationImplicit — Type

RungeKuttaToolKit.RKParameterization.RKParameterizationParallelExplicit — Type